Added key serialization support close #15, using own ASN1 definition stored in file - dependency libtasn1 - changed malloc/free functions, now using double pointer, dynamically allocates/frees objects - new module asn1 - some return codes checks fixed

+++ lib/src/bitpunch/math/mathctx.h
@@ -33,7 +33,7 @@ typedef struct _BPU_T_Math_Ctx {
 	BPU_T_GF2_16x *exp_table; ///< there are all elements referenced by i, so at i-th index is g^i element, g - generator
   	BPU_T_GF2_16x *log_table; ///< there are all indexes referenced by element, so alpha elemnet (g^i) -> i
   	BPU_T_GF2_16x mod; ///< polynomial modulus
-  	int mod_deg; ///< modulo degree, galois finite field GF(2^m)
+    uint8_t mod_deg; ///< modulo degree, galois finite field GF(2^m)
   	int ord; ///< group ord, number of elements
 }BPU_T_Math_Ctx;
